+++ examples/src/05_editing.c
@@ -75,10 +75,7 @@ int main(void) {
 
     // SubClassOf(pizza:Porcini ObjectAllValuesFrom(pizza:hasTopping
     // ObjectUnionOf(pizza:MozzarellaTopping pizza:PorciniTopping)))
-    UVec(CowlObjectPtr) vec = uvec(CowlObjectPtr);
-    uvec_push(CowlObjectPtr, &vec, mozzarella_topping);
-    uvec_push(CowlObjectPtr, &vec, porcini_topping);
-    CowlVector *operands = cowl_vector(&vec);
+    CowlVector *operands = cowl_vector_of(mozzarella_topping, porcini_topping);
     CowlNAryBool *closure = cowl_nary_bool(COWL_NT_UNION, operands);
     obj_quant = cowl_obj_quant(COWL_QT_ALL, has_topping, closure);
     axiom = cowl_sub_cls_axiom(porcini, obj_quant, NULL);